Combine Thought and Subthought components.
