--- src/components/Subthought.tsx
@@ -1,166 +1,0 @@
-import React, { useEffect, useMemo, useRef } from 'react'
-import { shallowEqual, useSelector } from 'react-redux'
-import { token } from '../../styled-system/tokens'
-import Autofocus from '../@types/Autofocus'
-import LazyEnv from '../@types/LazyEnv'
-import Path from '../@types/Path'
-import SimplePath from '../@types/SimplePath'
-import ThoughtId from '../@types/ThoughtId'
-import useChangeRef from '../hooks/useChangeRef'
-import attributeEquals from '../selectors/attributeEquals'
-import findFirstEnvContextWithZoom from '../selectors/findFirstEnvContextWithZoom'
-import { findAnyChild } from '../selectors/getChildren'
-import getContexts from '../selectors/getContexts'
-import getStyle from '../selectors/getStyle'
-import getThoughtById from '../selectors/getThoughtById'
-import isContextViewActive from '../selectors/isContextViewActive'
-import store from '../stores/app'
-import head from '../util/head'
-import isDescendantPath from '../util/isDescendantPath'
-import once from '../util/once'
-import NoOtherContexts from './NoOtherContexts'
-import Thought from './Thought'
-
-/** Renders a thought with style. */
-// TODO: These selectors can be optimized by calculating them once for all children, since they are the same among siblings. However siblings are not rendered contiguously (virtualTree), so they need to be calculated higher up.
-const Subthought = ({
-  autofocus,
-  debugIndex,
-  depth,
-  dropUncle,
-  env,
-  indexDescendant,
-  isMultiColumnTable,
-  leaf,
-  updateSize,
-  path,
-  prevChildId,
-  showContexts,
-  simplePath,
-  style,
-  zoomCursor,
-}: {
-  autofocus: Autofocus
-  debugIndex?: number
-  depth: number
-  dropUncle?: boolean
-  env?: LazyEnv
-  indexDescendant: number
-  isMultiColumnTable?: boolean
-  leaf?: boolean
-  updateSize?: () => void
-  path: Path
-  prevChildId?: ThoughtId
-  showContexts?: boolean
-  simplePath: SimplePath
-  style?: React.CSSProperties
-  zoomCursor?: boolean
-}) => {
-  const state = store.getState()
-  const ref = useRef<HTMLDivElement>(null)
-  const thought = useSelector(state => getThoughtById(state, head(simplePath)), shallowEqual)
-  const noOtherContexts = useSelector(
-    state => isContextViewActive(state, simplePath) && getContexts(state, thought.value).length <= 1,
-  )
-  const parentId = thought.parentId
-  const grandparentId = simplePath[simplePath.length - 3]
-  const isVisible = zoomCursor || autofocus === 'show' || autofocus === 'dim'
-  const autofocusChanged = useChangeRef(autofocus)
-
-  const childrenAttributeId = useSelector(
-    state =>
-      (thought.value !== '=children' && findAnyChild(state, parentId, child => child.value === '=children')?.id) ||
-      null,
-  )
-  const grandchildrenAttributeId = useSelector(
-    state =>
-      (thought.value !== '=style' &&
-        findAnyChild(state, grandparentId, child => child.value === '=grandchildren')?.id) ||
-      null,
-  )
-  const hideBullet = useSelector(state => {
-    const hideBulletsChildren = attributeEquals(state, childrenAttributeId, '=bullet', 'None')
-    if (hideBulletsChildren) return true
-    const hideBulletsGrandchildren =
-      thought.value !== '=bullet' && attributeEquals(state, grandchildrenAttributeId, '=bullet', 'None')
-    if (hideBulletsGrandchildren) return true
-    return false
-  })
-
-  /****************************/
-  const childEnvZoomId = once(() => findFirstEnvContextWithZoom(state, { id: thought.id, env }))
-
-  /** Returns true if the cursor is contained within the thought path, i.e. the thought is a descendant of the cursor. */
-  const isEditingChildPath = isDescendantPath(state.cursor, path)
-
-  const styleSelf = useMemo(
-    () => ({
-      ...(isEditingChildPath ? getStyle(state, childEnvZoomId()) : null),
-      ...style,
-    }),
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-    [isEditingChildPath, style],
-  )
-
-  // When autofocus changes, use a slow (750ms) ease-out to provide a gentle transition to non-focal thoughts.
-  // If autofocus has not changed, it means that the thought is being rendered for the first time, such as the children of a thought that was just expanded. In this case, match the tree-node top animation (150ms) to ensure that the newly rendered thoughts fade in to fill the space that is being opened up from the next uncle animating down.
-  // Note that ease-in is used in contrast to the tree-node's ease-out. This gives a little more time for the next uncle to animate down and clear space before the newly rendered thought fades in. Otherwise they overlap too much during the transition.
-  const opacity = autofocus === 'show' ? '1' : autofocus === 'dim' ? '0.5' : '0'
-  const opacityTransition = autofocusChanged
-    ? `opacity ${token('durations.layoutSlowShiftDuration')} ease-out`
-    : `opacity ${token('durations.layoutNodeAnimationDuration')} ease-in`
-  useEffect(() => {
-    if (!ref.current) return
-    // start opacity at 0 and set to actual opacity in useEffect
-    ref.current.style.opacity = opacity
-  })
-
-  // Short circuit if thought has already been removed.
-  // This can occur in a re-render even when thought is defined in the parent component.
-  if (!thought) return null
-
-  return (
-    <>
-      <div
-        ref={ref}
-        style={{
-          // Start opacity at 0 and set to actual opacity in useEffect.
-          // Do not fade in empty thoughts. An instant snap in feels better here.
-          // opacity creates a new stacking context, so it must only be applied to Thought, not to the outer VirtualThought which contains DropChild. Otherwise subsequent DropChild will be obscured.
-          opacity: thought.value === '' ? opacity : '0',
-          transition: opacityTransition,
-          pointerEvents: !isVisible ? 'none' : undefined,
-          // Safari has a known issue with subpixel calculations, especially during animations and with SVGs.
-          // This caused the thought to jerk slightly to the left at the end of the horizontal shift animation.
-          // By setting "will-change: transform;", we hint to the browser that the transform property will change in the future,
-          // allowing the browser to optimize the animation.
-          willChange: 'opacity',
-        }}
-      >
-        <Thought
-          debugIndex={debugIndex}
-          depth={depth + 1}
-          env={env}
-          hideBullet={hideBullet}
-          isContextPending={thought.value === '__PENDING__'}
-          leaf={leaf}
-          // isHeader={isHeader}
-          isHeader={false}
-          isMultiColumnTable={isMultiColumnTable}
-          isVisible={isVisible}
-          updateSize={updateSize}
-          path={path}
-          prevChildId={prevChildId}
-          rank={thought.rank}
-          showContexts={showContexts}
-          simplePath={simplePath}
-          style={styleSelf}
-        />
-      </div>
-
-      {noOtherContexts && <NoOtherContexts simplePath={simplePath} />}
-    </>
-  )
-}
-
-export default Subthought
