+++ src/components/Thought.tsx
@@ -1,7 +1,8 @@
 import classNames from 'classnames'
-import React, { useCallback, useMemo } from 'react'
+import React, { useCallback, useEffect, useMemo, useRef } from 'react'
 import { shallowEqual, useDispatch, useSelector } from 'react-redux'
 import { token } from '../../styled-system/tokens'
+import Autofocus from '../@types/Autofocus'
 import DragThoughtZone from '../@types/DragThoughtZone'
 import DropThoughtZone from '../@types/DropThoughtZone'
 import LazyEnv from '../@types/LazyEnv'
@@ -14,6 +15,7 @@ import { isTouch } from '../browser'
 import { AlertType, MAX_DISTANCE_FROM_CURSOR, REGEX_TAGS } from '../constants'
 import testFlags from '../e2e/testFlags'
 import globals from '../globals'
+import useChangeRef from '../hooks/useChangeRef'
 import useDragAndDropThought from '../hooks/useDragAndDropThought'
 import useDragHold from '../hooks/useDragHold'
 import useHideBullet from '../hooks/useHideBullet'
@@ -24,12 +26,16 @@ import attribute from '../selectors/attribute'
 import attributeEquals from '../selectors/attributeEquals'
 import childIdsToThoughts from '../selectors/childIdsToThoughts'
 import findDescendant from '../selectors/findDescendant'
+import findFirstEnvContextWithZoom from '../selectors/findFirstEnvContextWithZoom'
+import { findAnyChild } from '../selectors/getChildren'
 import { getAllChildrenAsThoughts, getChildrenRanked, hasChildren } from '../selectors/getChildren'
+import getContexts from '../selectors/getContexts'
 import getStyle from '../selectors/getStyle'
 import getThoughtById from '../selectors/getThoughtById'
 import isContextViewActive from '../selectors/isContextViewActive'
 import rootedParentOf from '../selectors/rootedParentOf'
 import themeColors from '../selectors/themeColors'
+import store from '../stores/app'
 import distractionFreeTypingStore from '../stores/distractionFreeTyping'
 import equalPath from '../util/equalPath'
 import equalThoughtRanked from '../util/equalThoughtRanked'
@@ -41,6 +47,7 @@ import isDescendantPath from '../util/isDescendantPath'
 import isDivider from '../util/isDivider'
 import isRoot from '../util/isRoot'
 import isURL from '../util/isURL'
+import once from '../util/once'
 import parentOf from '../util/parentOf'
 import publishMode from '../util/publishMode'
 import safeRefMerge from '../util/safeRefMerge'
@@ -48,6 +55,7 @@ import Bullet from './Bullet'
 import Byline from './Byline'
 import ContextBreadcrumbs from './ContextBreadcrumbs'
 import DropHover from './DropHover'
+import NoOtherContexts from './NoOtherContexts'
 import Note from './Note'
 import StaticThought from './StaticThought'
 
@@ -55,7 +63,7 @@ import StaticThought from './StaticThought'
  * Redux
  **********************************************************************/
 
-export interface ThoughtContainerProps {
+export interface ThoughtProps {
   allowSingleContext?: boolean
   childrenForced?: ThoughtId[]
   cursor?: Path | null
@@ -125,7 +133,7 @@ const ThoughtComponent = ({
   style: styleProp,
   styleContainer: styleContainerProp,
   updateSize,
-}: ThoughtContainerProps) => {
+}: ThoughtProps) => {
   const dispatch = useDispatch()
   const thoughtId = head(simplePath)
   const children = useSelector<Thought[]>(
@@ -492,4 +500,146 @@ const ThoughtComponent = ({
 ThoughtComponent.displayName = 'Thought'
 const ThoughtMemo = React.memo(ThoughtComponent)
 
-export default ThoughtMemo
+/** Renders a thought with style. */
+// TODO: These selectors can be optimized by calculating them once for all children, since they are the same among siblings. However siblings are not rendered contiguously (virtualTree), so they need to be calculated higher up.
+const ThoughtContainer = ({
+  autofocus,
+  debugIndex,
+  depth,
+  dropUncle,
+  env,
+  indexDescendant,
+  isMultiColumnTable,
+  leaf,
+  updateSize,
+  path,
+  prevChildId,
+  showContexts,
+  simplePath,
+  style,
+  zoomCursor,
+}: {
+  autofocus: Autofocus
+  debugIndex?: number
+  depth: number
+  dropUncle?: boolean
+  env?: LazyEnv
+  indexDescendant: number
+  isMultiColumnTable?: boolean
+  leaf?: boolean
+  updateSize?: () => void
+  path: Path
+  prevChildId?: ThoughtId
+  showContexts?: boolean
+  simplePath: SimplePath
+  style?: React.CSSProperties
+  zoomCursor?: boolean
+}) => {
+  const state = store.getState()
+  const ref = useRef<HTMLDivElement>(null)
+  const thought = useSelector(state => getThoughtById(state, head(simplePath)), shallowEqual)
+  const noOtherContexts = useSelector(
+    state => isContextViewActive(state, simplePath) && getContexts(state, thought.value).length <= 1,
+  )
+  const parentId = thought.parentId
+  const grandparentId = simplePath[simplePath.length - 3]
+  const isVisible = zoomCursor || autofocus === 'show' || autofocus === 'dim'
+  const autofocusChanged = useChangeRef(autofocus)
+
+  const childrenAttributeId = useSelector(
+    state =>
+      (thought.value !== '=children' && findAnyChild(state, parentId, child => child.value === '=children')?.id) ||
+      null,
+  )
+  const grandchildrenAttributeId = useSelector(
+    state =>
+      (thought.value !== '=style' &&
+        findAnyChild(state, grandparentId, child => child.value === '=grandchildren')?.id) ||
+      null,
+  )
+  const hideBullet = useSelector(state => {
+    const hideBulletsChildren = attributeEquals(state, childrenAttributeId, '=bullet', 'None')
+    if (hideBulletsChildren) return true
+    const hideBulletsGrandchildren =
+      thought.value !== '=bullet' && attributeEquals(state, grandchildrenAttributeId, '=bullet', 'None')
+    if (hideBulletsGrandchildren) return true
+    return false
+  })
+
+  /****************************/
+  const childEnvZoomId = once(() => findFirstEnvContextWithZoom(state, { id: thought.id, env }))
+
+  /** Returns true if the cursor is contained within the thought path, i.e. the thought is a descendant of the cursor. */
+  const isEditingChildPath = isDescendantPath(state.cursor, path)
+
+  const styleSelf = useMemo(
+    () => ({
+      ...(isEditingChildPath ? getStyle(state, childEnvZoomId()) : null),
+      ...style,
+    }),
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+    [isEditingChildPath, style],
+  )
+
+  // When autofocus changes, use a slow (750ms) ease-out to provide a gentle transition to non-focal thoughts.
+  // If autofocus has not changed, it means that the thought is being rendered for the first time, such as the children of a thought that was just expanded. In this case, match the tree-node top animation (150ms) to ensure that the newly rendered thoughts fade in to fill the space that is being opened up from the next uncle animating down.
+  // Note that ease-in is used in contrast to the tree-node's ease-out. This gives a little more time for the next uncle to animate down and clear space before the newly rendered thought fades in. Otherwise they overlap too much during the transition.
+  const opacity = autofocus === 'show' ? '1' : autofocus === 'dim' ? '0.5' : '0'
+  const opacityTransition = autofocusChanged
+    ? `opacity ${token('durations.layoutSlowShiftDuration')} ease-out`
+    : `opacity ${token('durations.layoutNodeAnimationDuration')} ease-in`
+  useEffect(() => {
+    if (!ref.current) return
+    // start opacity at 0 and set to actual opacity in useEffect
+    ref.current.style.opacity = opacity
+  })
+
+  // Short circuit if thought has already been removed.
+  // This can occur in a re-render even when thought is defined in the parent component.
+  if (!thought) return null
+
+  return (
+    <>
+      <div
+        ref={ref}
+        style={{
+          // Start opacity at 0 and set to actual opacity in useEffect.
+          // Do not fade in empty thoughts. An instant snap in feels better here.
+          // opacity creates a new stacking context, so it must only be applied to Thought, not to the outer VirtualThought which contains DropChild. Otherwise subsequent DropChild will be obscured.
+          opacity: thought.value === '' ? opacity : '0',
+          transition: opacityTransition,
+          pointerEvents: !isVisible ? 'none' : undefined,
+          // Safari has a known issue with subpixel calculations, especially during animations and with SVGs.
+          // This caused the thought to jerk slightly to the left at the end of the horizontal shift animation.
+          // By setting "will-change: transform;", we hint to the browser that the transform property will change in the future,
+          // allowing the browser to optimize the animation.
+          willChange: 'opacity',
+        }}
+      >
+        <ThoughtMemo
+          debugIndex={debugIndex}
+          depth={depth + 1}
+          env={env}
+          hideBullet={hideBullet}
+          isContextPending={thought.value === '__PENDING__'}
+          leaf={leaf}
+          // isHeader={isHeader}
+          isHeader={false}
+          isMultiColumnTable={isMultiColumnTable}
+          isVisible={isVisible}
+          updateSize={updateSize}
+          path={path}
+          prevChildId={prevChildId}
+          rank={thought.rank}
+          showContexts={showContexts}
+          simplePath={simplePath}
+          style={styleSelf}
+        />
+      </div>
+
+      {noOtherContexts && <NoOtherContexts simplePath={simplePath} />}
+    </>
+  )
+}
+
+export default ThoughtContainer
